+++ guice/src/main/java/cucumber/runtime/java/guice/impl/InjectorSourceFactory.java
@@ -3,23 +3,22 @@ package cucumber.runtime.java.guice.impl;
 import com.google.inject.Guice;
 import com.google.inject.Injector;
 import com.google.inject.Stage;
+import cucumber.runtime.Env;
 import cucumber.runtime.java.guice.InjectorSource;
-
-import java.util.Properties;
 
 import static java.text.MessageFormat.format;
 
 public class InjectorSourceFactory {
 
     public static final String GUICE_INJECTOR_SOURCE_KEY = "guice.injector-source";
-    private final Properties properties;
+    private final Env env;
 
-    public InjectorSourceFactory(Properties properties) {
-        this.properties = properties;
+    public InjectorSourceFactory(Env env) {
+        this.env = env;
     }
 
     public InjectorSource create() {
-        String injectorSourceClassName = properties.getProperty(GUICE_INJECTOR_SOURCE_KEY);
+        String injectorSourceClassName = env.get(GUICE_INJECTOR_SOURCE_KEY);
         if (injectorSourceClassName == null) {
             return createDefaultScenarioModuleInjectorSource();
         } else {
@@ -41,7 +40,7 @@ public class InjectorSourceFactory {
         try {
             return (InjectorSource) Class.forName(injectorSourceClassName).newInstance();
         } catch (Exception e) {
-            String message = format("Instantiation of ''{0}'' failed. Checked the caused by exception and ensure your" +
+            String message = format("Instantiation of ''{0}'' failed. Check the caused by exception and ensure your" +
                     "InjectorSource implementation is accessible and has a public zero args constructor.",
                     injectorSourceClassName);
             throw new InjectorSourceInstantiationFailed(message, e);