+++ guice/src/test/java/cucumber/runtime/java/guice/impl/InjectorSourceFactoryTest.java
@@ -1,8 +1,8 @@
 package cucumber.runtime.java.guice.impl;
 
 import com.google.inject.Injector;
+import cucumber.runtime.Env;
 import cucumber.runtime.java.guice.InjectorSource;
-import org.junit.Before;
 import org.junit.Test;
 
 import java.util.Properties;
@@ -13,33 +13,37 @@ import static org.junit.Assert.assertThat;
 import static org.junit.Assert.fail;
 
 public class InjectorSourceFactoryTest {
-    private Properties properties;
-    private InjectorSourceFactory injectorSourceFactory;
 
-    @Before
-    public void setUp() {
-        properties = new Properties();
-        injectorSourceFactory = new InjectorSourceFactory(properties);
+    private InjectorSourceFactory createInjectorSourceFactory(Properties properties) {
+        return new InjectorSourceFactory(new Env(properties));
     }
 
     @Test
     public void createsDefaultInjectorSourceWhenGuiceModulePropertyIsNotSet() throws Exception {
+        InjectorSourceFactory injectorSourceFactory = createInjectorSourceFactory(new Properties());
         assertThat(injectorSourceFactory.create(), is(instanceOf(InjectorSource.class)));
     }
 
     static class CustomInjectorSource implements InjectorSource {
-        @Override public Injector getInjector() { return null; }
+        @Override
+        public Injector getInjector() {
+            return null;
+        }
     }
 
     @Test
     public void instantiatesInjectorSourceByFullyQualifiedName() throws Exception {
-        properties.setProperty(InjectorSourceFactory.GUICE_INJECTOR_SOURCE_KEY, nameOf(CustomInjectorSource.class));
+        Properties properties = new Properties();
+        properties.setProperty(InjectorSourceFactory.GUICE_INJECTOR_SOURCE_KEY, CustomInjectorSource.class.getName());
+        InjectorSourceFactory injectorSourceFactory = createInjectorSourceFactory(properties);
         assertThat(injectorSourceFactory.create(), is(instanceOf(CustomInjectorSource.class)));
     }
 
     @Test
     public void failsToInstantiateNonExistantClass() throws Exception {
+        Properties properties = new Properties();
         properties.setProperty(InjectorSourceFactory.GUICE_INJECTOR_SOURCE_KEY, "some.bogus.Class");
+        InjectorSourceFactory injectorSourceFactory = createInjectorSourceFactory(properties);
         try {
             injectorSourceFactory.create();
             fail();
@@ -50,7 +54,9 @@ public class InjectorSourceFactoryTest {
 
     @Test
     public void failsToInstantiateClassNotImplementingInjectorSource() throws Exception {
-        properties.setProperty(InjectorSourceFactory.GUICE_INJECTOR_SOURCE_KEY, nameOf(String.class));
+        Properties properties = new Properties();
+        properties.setProperty(InjectorSourceFactory.GUICE_INJECTOR_SOURCE_KEY, String.class.getName());
+        InjectorSourceFactory injectorSourceFactory = createInjectorSourceFactory(properties);
         try {
             injectorSourceFactory.create();
             fail();
@@ -60,13 +66,20 @@ public class InjectorSourceFactoryTest {
     }
 
     static class PrivateConstructor implements InjectorSource {
-        private PrivateConstructor() {}
-        @Override public Injector getInjector() { return null; }
+        private PrivateConstructor() {
+        }
+
+        @Override
+        public Injector getInjector() {
+            return null;
+        }
     }
 
     @Test
     public void failsToInstantiateClassWithPrivateConstructor() throws Exception {
-        properties.setProperty(InjectorSourceFactory.GUICE_INJECTOR_SOURCE_KEY, nameOf(PrivateConstructor.class));
+        Properties properties = new Properties();
+        properties.setProperty(InjectorSourceFactory.GUICE_INJECTOR_SOURCE_KEY, PrivateConstructor.class.getName());
+        InjectorSourceFactory injectorSourceFactory = createInjectorSourceFactory(properties);
         try {
             injectorSourceFactory.create();
             fail();
@@ -76,13 +89,20 @@ public class InjectorSourceFactoryTest {
     }
 
     static class NoDefaultConstructor implements InjectorSource {
-        private NoDefaultConstructor(String someParameter) {}
-        @Override public Injector getInjector() { return null; }
+        private NoDefaultConstructor(String someParameter) {
+        }
+
+        @Override
+        public Injector getInjector() {
+            return null;
+        }
     }
 
     @Test
     public void failsToInstantiateClassWithNoDefaultConstructor() throws Exception {
-        properties.setProperty(InjectorSourceFactory.GUICE_INJECTOR_SOURCE_KEY, nameOf(NoDefaultConstructor.class));
+        Properties properties = new Properties();
+        properties.setProperty(InjectorSourceFactory.GUICE_INJECTOR_SOURCE_KEY, NoDefaultConstructor.class.getName());
+        InjectorSourceFactory injectorSourceFactory = createInjectorSourceFactory(properties);
         try {
             injectorSourceFactory.create();
             fail();
@@ -90,10 +110,4 @@ public class InjectorSourceFactoryTest {
             assertThat(exception.getCause(), instanceOf(InstantiationException.class));
         }
     }
-
-    private String nameOf(Class<?> moduleClass) {
-        return moduleClass.getName();
-    }
-
-
 }